getting the dao to work

+++ newcomer/src/main/java/com/otms/newcomer/dao/OrderDao.java
@@ -2,6 +2,6 @@ package com.otms.newcomer.dao;
 
 import com.otms.newcomer.entity.Order;
 
-public interface OrderDao extends GenericDao<Order, Integer> {
+public interface OrderDao extends GenericDao<Order, Long> {
 
 }
